Use Access Credentials instead of Authorized Client.
Since the client cannot be serialized into core events, it cannot be
passed to signal receivers as in the original polymer.js GoogleSignIn
implementation. Thus we are using the Access Credential to instantiate
new clients where necessary. This makes the use of the client obsolete,
therefore using the more simple access credential flow is preferred.